refactor: collapse the opt-outs to CONFIG_DISABLE_OTA and CONFIG_DISABLE_SMS
Two states are wanted: SMS without RFM, and SMS with RFM. REFRESH is only
ever fed by remote commands, so it goes with OTA; SMS off implies OTA off;
CAT stays in every build so the terminal sees the same card at boot. A
compiled-out subsystem logs the profile fix when the terminal still offers
it, and tests/opt_out pins the gated answers.
