+++ CMakeLists.txt
@@ -20,10 +20,8 @@ option(CONFIG_COMPACT_STORAGE 	"Build with compact storage implementation")
 option(CONFIG_BUILD_LIB_ONLY 	"Build libraries only")
 option(CONFIG_ALT_FILE_SEPARATOR "Use _ as file separator instead of /")
 option(CONFIG_BUILD_FUZZERS 	"Build libFuzzer targets in tests/fuzz (clang only, implies sanitizers)")
-option(CONFIG_DISABLE_OTA 		"Compile out OTA/RFM remote command handling")
-option(CONFIG_DISABLE_SMS 		"Compile out SMS handling (requires CONFIG_DISABLE_OTA)")
-option(CONFIG_DISABLE_REFRESH 	"Compile out REFRESH file-change tracking")
-option(CONFIG_DISABLE_PROACTIVE "Compile out proactive/CAT commands (requires CONFIG_DISABLE_SMS and CONFIG_DISABLE_REFRESH)")
+option(CONFIG_DISABLE_OTA 		"Compile out OTA/RFM remote command handling and the REFRESH it drives")
+option(CONFIG_DISABLE_SMS 		"Compile out SMS handling (implies CONFIG_DISABLE_OTA)")
 
 set(CONFIG_SS_STORAGE_PATH_DEFAULT "./files" CACHE STRING "Default storage path for SoftSIM filesystem")
 set(CONFIG_SS_STORAGE_PATH_MAX "100" CACHE STRING "Maximum path length for storage operations")
@@ -66,15 +64,11 @@ if(CONFIG_EXTERNAL_KEY_LOAD)
 	add_compile_definitions(CONFIG_EXTERNAL_KEY_LOAD)
 endif()
 
-# The opt-outs mirror the reachability graph: OTA is reached only through SMS,
-# SMS and REFRESH only through the proactive machinery (see context.h).
-if(CONFIG_DISABLE_SMS AND NOT CONFIG_DISABLE_OTA)
-	message(FATAL_ERROR "CONFIG_DISABLE_SMS requires CONFIG_DISABLE_OTA (OTA is delivered over SMS)")
+# OTA is delivered over SMS, so compiling out SMS compiles out OTA as well.
+if(CONFIG_DISABLE_SMS)
+	set(CONFIG_DISABLE_OTA ON)
 endif()
-if(CONFIG_DISABLE_PROACTIVE AND NOT (CONFIG_DISABLE_SMS AND CONFIG_DISABLE_REFRESH))
-	message(FATAL_ERROR "CONFIG_DISABLE_PROACTIVE requires CONFIG_DISABLE_SMS and CONFIG_DISABLE_REFRESH")
-endif()
-foreach(opt CONFIG_DISABLE_OTA CONFIG_DISABLE_SMS CONFIG_DISABLE_REFRESH CONFIG_DISABLE_PROACTIVE)
+foreach(opt CONFIG_DISABLE_OTA CONFIG_DISABLE_SMS)
 	if(${opt})
 		add_compile_definitions(${opt})
 	endif()
